+++ packages/react-mui-store-file-viewer/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.66] - 2024-09-16
+
+### Fixed
+
+- "Get Link" password method helper text corrected
+
+### Added
+
+- Show modal with copy password dialog after "Share via Email" password link creation
+
 ## [0.0.65] - 2024-09-04
 
 ### Added